🐛 Move misplaced types file :sweat_smile:
